Applied #231 to flow generation and extracted more common logic
